third_party/blink/web_tests/platform/generic/http/tests/devtools/tracing/timeline-style/timeline-style-recalc-with-invalidator-invalidations-expected.txt

Tests the Timeline API instrumentation of style recalc events with invalidations.


Running: testLocalFrame
first recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:29}
        changedAttribute : undefined
        changedClass : "red"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:30}
        changedAttribute : undefined
        changedClass : "red"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='red'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: multipleStyleRecalcs
first recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:40}
        changedAttribute : undefined
        changedClass : "green"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='green'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:40}
        changedAttribute : undefined
        changedClass : "red"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='green'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
second recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:42}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:42}
        changedAttribute : undefined
        changedClass : "green"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:43}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:43}
        changedAttribute : undefined
        changedClass : "red"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
third recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:45}
        changedAttribute : undefined
        changedClass : "snow"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:45}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:46}
        changedAttribute : undefined
        changedClass : "snow"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:46}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:47}
        changedAttribute : undefined
        changedClass : "snow"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementThree' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: testSubframe
first recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:58}
        changedAttribute : undefined
        changedClass : "green"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='green'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
second recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:60}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:60}
        changedAttribute : undefined
        changedClass : "green"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:61}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='blue'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
third recalculate styles[
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:63}
        changedAttribute : undefined
        changedClass : "snow"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:63}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementOne' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:64}
        changedAttribute : undefined
        changedClass : "snow"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:64}
        changedAttribute : undefined
        changedClass : "blue"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementTwo' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Element has pending invalidation list, stackTrace: test://evaluations/0/timeline-style-recalc-with-invalidator-invalidations.js:65}
        changedAttribute : undefined
        changedClass : "snow"
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElementThree' class='snow'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
